refactor: simplify mobile web auth and provider request internals
Share zod issue flattening between Remote and ProviderRequest, memoize
the wata host discovery document per adapter instance, and drop dead
code from parseMetadata and the localnet test host.
